requirements.txt
absl-py==0.9.0
appdirs==1.4.3
appnope==0.1.0
astroid==2.3.3
astunparse==1.6.3
attrs==19.3.0
autopep8==1.5
backcall==0.1.0
black==19.10b0
bleach==3.1.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
chart-studio==1.1.0
click==7.1.1
colorama==0.4.3
colorlover==0.3.0
configparser==5.0.0
cufflinks==0.17.3
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
docker-pycreds==0.4.0
entrypoints==0.3
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
future==0.18.2
gast==0.3.3
gitdb==4.0.4
GitPython==3.1.1
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gql==0.2.0
graphql-core==1.1
grpcio==1.28.1
h5py==2.10.0
idna==2.9
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
joblib==0.14.1
json5==0.8.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.5
jupyterlab-server==1.0.6
Keras-Preprocessing==1.1.0
keras-tuner==1.0.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mlxtend==0.17.2
nbconvert==5.6.1
nbformat==5.0.3
notebook==6.0.2
numpy==1.18.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opt-einsum==3.2.0
pandas==1.0.0
pandocfilters==1.4.2
parso==0.5.2
pathspec==0.7.0
pathtools==0.1.2
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
plotly==4.6.0
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.2
protobuf==3.11.3
psutil==5.7.0
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.7
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
PyYAML==5.3.1
pyzmq==18.1.1
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
sentry-sdk==0.14.3
shortuuid==1.0.1
six==1.14.0
smmap==3.0.2
spotipy==2.12.0
SQLAlchemy==1.3.13
statsmodels==0.11.1
subprocess32==3.5.4
tabulate==0.8.7
tensorboard==2.2.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0rc3
tensorflow-estimator==2.2.0rc0
termcolor==1.1.0
terminado==0.8.3
terminaltables==3.1.0
testpath==0.4.4
toml==0.10.0
tornado==6.0.3
tqdm==4.45.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==16.7.9
wandb==0.8.31
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
